backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pymongo==4.1.1
python-dotenv==0.19.0
ijson==3.1.4
bson==0.5.10